move ideal gas models to feos
