+++ src/python/cubic.rs
@@ -1,5 +1,5 @@
 use feos_core::python::cubic::*;
-use feos_core::python::parameter::*;
+use feos_core::python::parameter::{PyChemicalRecord, PyIdentifier};
 use pyo3::prelude::*;
 
 #[pymodule]
